AmountCell component
