+++ src/components/TokenOptions.tsx
@@ -128,7 +128,9 @@ export function TokenOptions({
                 ) : (
                   <Eye className='mr-2 h-4 w-4' aria-hidden='true' />
                 )}
-                {showHiddenCats ? t`Conceal Hidden Tokens` : t`Reveal Hidden Tokens`}
+                {showHiddenCats
+                  ? t`Conceal Hidden Tokens`
+                  : t`Reveal Hidden Tokens`}
               </DropdownMenuItem>
               <DropdownMenuItem
                 className='cursor-pointer'
